feature:[BA-22] Reworked table column settings
