+++ cfx-frontend/src/app/modules/shared/components/parts-table/parts-table.component.scss
@@ -31,6 +31,15 @@
   cursor: pointer;
 }
 
+.close-settings-icon {
+  margin: 0;
+  margin-left: 5px;
+  width: 14px;
+  height: 14px;
+  padding: 0px;
+  cursor: pointer;
+}
+
 .table-menu {
   width: 40px;
 }
@@ -376,4 +385,30 @@ tr.error {
 
 .table--body--row .table--cell--sticky {
   padding: 0px !important;
+}
+
+.menu-button {
+  padding: 10px;
+  display: flex;
+}
+
+.menu-text {
+  @extend .body-medium;
+  font-weight: 400;
+  font-size: 18px;
+  line-height: 25.2px;
+  padding-left: 10px;
+}
+
+.menu-header {
+  @extend .title-medium;
+  color: #333333;
+  font-weight: 600;
+  font-size: 20px;
+  line-height: 24px;
+  padding: 0 10px 10px 10px;
+}
+
+.divider {
+  color: #d2d2d2;
 }
