+++ cfx-frontend/src/app/modules/shared/components/table-settings/table-settings.component.html
@@ -16,78 +16,36 @@ under the License.
 
 SPDX-License-Identifier: Apache-2.0
 -->
-<div class="container"
-  cdkDropListGroup>
-  <div class="dialog--content">
-    <div class="dialog--header--container">
-      <div class="dialog--header--text">{{title | i18n}}</div>
-      <div>
-        <img class="close-button"
-          (click)="dialogRef.close()"
-          src="/assets/images/icons/closeDialog_icon.svg"
-          width="14px"
-          height="14px"
-          alt="X" />
-      </div>
-    </div>
-    <div class="dialog--content--container">
-      <div class="dialog--columns--container">
-        <div class="select-all-container">
-          <mat-checkbox color="accent"
-            class="dialog--columns--columns--list"
-            (change)="selectAll($event.checked)"
-            [checked]="selectAllSelected">{{'table.tableSettings.selectAll' | i18n}}
-          </mat-checkbox>
-          <img matTooltip="{{'table.tableSettings.refreshTooltip' | i18n}}"
-            matTooltipClass="table--header--tooltip"
-            matTooltipPosition="above"
-            class="close-button"
-            [class.mdc-tooltip--multiline]="true"
-            [matTooltipShowDelay]="1000"
-            (click)="resetColumns()"
-            src="/assets/images/icons/refresh_icon.svg"
-            width="17.67px"
-            height="15px"
-            alt="<" />
-        </div>
-        <mat-divider class="divider"></mat-divider>
-        <div cdkDropList
-          [cdkDropListData]="dialogColumns"
-          (cdkDropListDropped)="drop($event)">
-          <ng-container class="dialog--columns--columns--list"
-            *ngFor="let item of dialogColumns; let index=index">
-            <div *ngIf="item !== 'select' && item !== 'menu' && item !== 'settings'"
-              [ngClass]="dragActive ? 'dialog--item--container_static' : 'dialog--item--container'"
-              (cdkDragStarted)="dragging(true)"
-              (cdkDragEnded)="dragging(false)"
-              cdkDrag>
-              <img class="drag-icon"
-                src="/assets/images/drag_indicator.svg"
-                width="8.46px"
-                height="14.46px"
-                alt="↓" />
-              <mat-checkbox color="accent"
-                [checked]="columnOptions.get(item)"
-                [value]="item"
-                [class.selected-item]="selectedColumn === item"
-                (change)="$event ? handleCheckBoxChange(item, $event.checked) : null"
-                (click)="handleListItemClick($event, item)"
-                class="dialog--columns--columns--list">
-                {{'table.column.' + item | i18n}}
-                <div class="placeholder"
-                  *cdkDragPlaceholder></div>
-              </mat-checkbox>
-            </div>
-          </ng-container>
-        </div>
-      </div>
-    </div>
+<div class="menu-header">{{ 'table.tableSettings.titleSettings' | i18n }}</div>
+<button mat-menu-item>
+  <div class="menu-button"
+    (click)="resetFilters()">
+    <img matTooltip="{{'table.tableSettings.refreshTooltip' | i18n}}"
+      matTooltipClass="table--header--tooltip"
+      matTooltipPosition="above"
+      class="reset-button"
+      [class.mdc-tooltip--multiline]="true"
+      [matTooltipShowDelay]="1000"
+      src="/assets/images/icons/reset_filters_icon.svg"
+      width="16px"
+      height="18.45px"
+      alt="<" />
+    <div class="menu-text">{{ 'table.tableSettings.resetAllFilters' | i18n }}</div>
   </div>
-
-  <!-- <mat-dialog-actions class="dialog--actions--container">
-    <div class="dialog--actions--container">
-      <button class="mat-raised-button mat-primary dialog--actions--save--button"
-        (click)="save()">{{'table.tableSettings.saveAction' | i18n }}</button>
-    </div>
-  </mat-dialog-actions> -->
-</div>
+</button>
+<mat-divider class="divider"></mat-divider>
+<button mat-menu-item>
+  <div class="menu-button"
+    (click)="openTableColumnSettingsDialog()">
+    <img matTooltip="{{'table.tableSettings.refreshTooltip' | i18n}}"
+      matTooltipClass="table--header--tooltip"
+      matTooltipPosition="above"
+      [class.mdc-tooltip--multiline]="true"
+      [matTooltipShowDelay]="1000"
+      src="/assets/images/icons/column_settings_icon.svg"
+      width="20.1px"
+      height="20.03px"
+      alt="<" />
+    <div class="menu-text">{{ 'table.tableSettings.title' | i18n }}</div>
+  </div>
+</button>
