+++ cfx-frontend/src/app/modules/shared/components/table-settings/table-settings.component.scss
@@ -1,165 +1,30 @@
-/********************************************************************************
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ********************************************************************************/
 @import '../../../../../theme/base.scss';
 @tailwind base;
 
-.container {
-  width: 200px;
-  overflow: hidden !important;
+.menu-button {
+  padding: 10px;
+  display: flex;
 }
 
-.dialog--item--container {
-  width: 100%;
-  padding: 3px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  box-sizing: border-box;
+.menu-text {
+  @extend .body-medium;
+  font-weight: 400;
+  font-size: 18px;
+  line-height: 25.2px;
+  padding-left: 10px;
 }
 
-.dialog--item--container_static {
-  width: 100%;
-  padding: 3px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  box-sizing: border-box;
+.menu-header {
+  @extend .title-medium;
+  color: #333333;
+  font-weight: 600;
+  font-size: 20px;
+  line-height: 24px;
+  padding: 0 10px 10px 10px;
 }
 
-.dialog--content {
-  display: flex;
-  flex-direction: column;
-
-  .select-all-container {
-    display: flex;
-    justify-content: space-between;
-    padding-left: 21.5px;
-  }
-
-  .placeholder {
-    background: #ccc;
-    border: dotted 3px #999;
-    min-height: 24px;
-    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
-  }
-
-  .dialog--header--container {
-    display: flex;
-    padding-bottom: 24px;
-    justify-content: space-between;
-  }
-
-  .mat-icon {
-    width: 32px;
-    height: 32px;
-    font-size: 30px;
-    margin-top: 4px;
-    cursor: pointer;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .dialog--content--menu--order {
-    display: flex;
-    gap: 8px;
-    margin-right: 12px;
-  }
-
-  .dialog--header--text {
-    @extend .title-small;
-    font-weight: 600;
-    font-size: 16px;
-    text-align: left;
-    justify-self: center;
-    color: black;
-  }
-
-  .dialog--columns--container {
-    font-size: 18px;
-  }
-
-  .drag-icon {
-    margin-left: 5px;
-    margin-right: 5px;
-    opacity: 0;
-  }
-
-  .dialog--item--container:hover {
-    background-color: rgb(237, 240, 244);
-  }
-
-  .dialog--item--container:hover img {
-    opacity: 1;
-  }
-
-  .dialog--columns--columns--list {
-    @extend .body-xsmall;
-    font-size: 12px;
-    line-height: 13.2px;
-    font-weight: 400;
-  }
-}
-
-.dialog--actions--container {
-  background-color: rgb(237, 240, 244);
-  width: 100%;
-  display: flex;
-  justify-content: center;
-
-  .dialog--actions--save--button {
-    display: inline-flex;
-    -webkit-box-align: center;
-    align-items: center;
-    -webkit-box-pack: center;
-    justify-content: center;
-    position: relative;
-    box-sizing: border-box;
-    -webkit-tap-highlight-color: transparent;
-    outline: 0px;
-    border: 0px;
-    margin: 0px;
-    cursor: pointer;
-    user-select: none;
-    vertical-align: middle;
-    appearance: none;
-    text-decoration: none;
-    font-family: LibreFranklin, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial,
-      sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol';
-    font-weight: 500;
-    line-height: 1.5;
-    color: rgb(255, 255, 255);
-    min-width: 64px;
-    transition: background-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,
-      box-shadow 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms, border-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,
-      color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-
-    background-color: rgb(15, 113, 203);
-    border-radius: 50px;
-    font-size: 18px;
-    padding: 14px 32px;
-    box-shadow: rgba(15, 113, 203, 0.4) 0px 0px 0px 3px;
-  }
-
-  .dialog--actions--save--button:hover,
-  .dialog--actions--save--button:active,
-  .dialog--actions--save--button:focus {
-    background-color: rgb(13, 85, 175);
-  }
+.divider {
+  color: #d2d2d2;
+  margin-left: 10px;
+  margin-right: 10px;
 }
