+++ cfx-frontend/src/app/modules/shared/components/table-settings/table-settings.component.ts
@@ -1,176 +1,37 @@
-/********************************************************************************
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ********************************************************************************/
-
-import { Component, EventEmitter, Inject, Output, } from '@angular/core';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { TableViewSettings } from '@core/user/table-settings.model';
-import { TableSettingsService } from '@core/user/table-settings.service';
-import {
-  CdkDragDrop,
-  moveItemInArray,
-  transferArrayItem,
-} from '@angular/cdk/drag-drop';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { MatDialogConfig, MatDialog } from '@angular/material/dialog';
+import { TableColumnSettingsComponent } from '../table-column-settings/table-column-settings.component';
 import { TableType } from '../multi-select-autocomplete/table-type.model';
 
 @Component({
   selector: 'app-table-settings',
-  templateUrl: 'table-settings.component.html',
-  styleUrls: ['table-settings.component.scss'],
+  templateUrl: './table-settings.component.html',
+  styleUrls: ['./table-settings.component.scss']
 })
 export class TableSettingsComponent {
 
-  @Output() changeSettingsEvent = new EventEmitter<void>();
-  title: string;
-  panelClass: string;
+  @Input() tableType: TableType;
+  @Input() defaultColumns: string[];
+  @Input() defaultFilterColumns: string[];
 
-  tableType: TableType;
-  defaultColumns: string[];
-  defaultFilterColumns: string[];
+  @Output() filtersReset = new EventEmitter<void>();
 
-  columnOptions: Map<string, boolean>;
-  dialogColumns: string[];
-  tableColumns: string[];
-  filterColumns: string[];
+  constructor(private dialog: MatDialog) { }
 
-  selectAllSelected: boolean;
-  selectedColumn: string = null;
-
-  isCustomerTable: boolean;
-
-  dragActive = false;
-
-  constructor(public dialogRef: MatDialogRef<TableSettingsComponent>, @Inject(MAT_DIALOG_DATA) public data: any, public readonly tableSettingsService: TableSettingsService) {
-    // Layout
-    this.title = data.title;
-    this.panelClass = data.panelClass;
-    this.isCustomerTable = data.tableType === TableType.AS_BUILT_CUSTOMER || data.tableType === TableType.AS_PLANNED_CUSTOMER;
-    // Passed Data
-    this.tableType = data.tableType;
-    this.defaultColumns = data.defaultColumns;
-    this.defaultFilterColumns = data.defaultFilterColumns;
-
-    this.loadSettings(tableSettingsService, data);
-
-    if (dialogRef?.afterClosed)
-      dialogRef.afterClosed().subscribe(() => {
-        this.save();
-      });
+  resetFilters(): void {
+    this.filtersReset.emit();
   }
 
-  async loadSettings(tableSettingsService: TableSettingsService, data: any) {
-    const storedSettings = await tableSettingsService.getStoredTableSettings()?.[this.tableType];
-    if (storedSettings) {
-      // Storage Data
-      this.columnOptions = storedSettings.columnSettingsOptions;
-      this.dialogColumns = storedSettings.columnsForDialog;
-      this.tableColumns = storedSettings.columnsForTable;
-      this.filterColumns = storedSettings.filterColumnsForTable;
-    } else {
-      this.dialogColumns = data.defaultColumns;
-      this.tableColumns = data.defaultColumns;
-      this.filterColumns = data.defaultColumns;
-      this.columnOptions = new Map<string, boolean>();
-      this.selectAll(true);
-    }
-
-    this.selectAllSelected = this.dialogColumns.length === this.tableColumns.length;
-  }
-
-  async save() {
-    // build new tableColumns how they should be displayed
-    const newTableColumns: string[] = [];
-    const newTableFilterColumns: string[] = [];
-    // iterate over dialogColumns
-    for (const column of this.dialogColumns) {
-      // if item in dialogColumns is true in columnOptions --> add to new tableColumns
-      if (this.columnOptions.get(column) || column === 'select' || column === 'menu' || column === 'settings') {
-        newTableColumns.push(column);
-        // ignore select column in customertable
-        if ((column === 'select') && !this.isCustomerTable) {
-          newTableFilterColumns.push('Filter');
-        } else {
-          newTableFilterColumns.push('filter' + column.charAt(0).toUpperCase() + column.slice(1));
-        }
-      }
-    }
-
-    // get Settingslist
-    const tableSettingsList = await this.tableSettingsService.getStoredTableSettings();
-
-    // set this tableType Settings from SettingsList to the new one
-    tableSettingsList[this.tableType] = {
-      columnSettingsOptions: this.columnOptions,
-      columnsForDialog: this.dialogColumns,
-      columnsForTable: newTableColumns,
-      filterColumnsForTable: newTableFilterColumns
-    } as TableViewSettings;
-
-    // save all values back to localstorage
-    this.tableSettingsService.storeTableSettings(tableSettingsList);
-
-    // trigger action that table will refresh
-    this.tableSettingsService.emitChangeEvent();
-  }
-
-  handleCheckBoxChange(item: string, isChecked: boolean) {
-    this.columnOptions.set(item, isChecked);
-  }
-
-  handleListItemClick(event: MouseEvent, item: string) {
-    const element = event.target as HTMLElement;
-
-    if (element.tagName !== 'INPUT') {
-      this.selectedColumn = item;
-      element.classList.toggle('selected-item');
-    }
-  }
-
-  dragging(state: boolean) {
-    this.dragActive = state;
-  }
-
-  selectAll(isChecked: boolean) {
-    for (const column of this.dialogColumns) {
-      if (column === 'select' || column === 'menu' || column === 'settings') {
-        continue;
-      }
-      this.columnOptions.set(column, isChecked);
-    }
-    this.selectAllSelected = true;
-  }
-
-  resetColumns() {
-    this.dialogColumns = [...this.defaultColumns];
-    this.selectAll(true);
-  }
-
-  drop(event: CdkDragDrop<string[]>) {
-    const offset = event.container.data.includes('select') ? 1 : 0;
-    if (event.previousContainer === event.container) {
-      moveItemInArray(event.container.data, event.previousIndex + offset, event.currentIndex + offset);
-    } else {
-      transferArrayItem(
-        event.previousContainer.data,
-        event.container.data,
-        event.previousIndex,
-        event.currentIndex,
-      );
-    }
+  openTableColumnSettingsDialog(): void {
+    const config = new MatDialogConfig();
+    config.autoFocus = false;
+    config.panelClass = 'table-column-settings-dialog';
+    config.data = {
+      title: 'table.tableSettings.title',
+      tableType: this.tableType,
+      defaultColumns: this.defaultColumns,
+      defaultFilterColumns: this.defaultFilterColumns
+    };
+    this.dialog.open(TableColumnSettingsComponent, config);
   }
 }
