+++ cfx-frontend/src/app/modules/shared/components/table/table.component.scss
@@ -41,6 +41,15 @@ table {
   cursor: pointer;
 }
 
+.close-settings-icon {
+  margin: 0;
+  margin-left: 5px;
+  width: 14px;
+  height: 14px;
+  padding: 0px;
+  cursor: pointer;
+}
+
 .notification-table--select-all {
   display: inline-flex;
   align-items: center;
