+++ cfx-frontend/src/app/modules/shared/modules/part-details/presentation/start-investigation/start-investigation.component.spec.ts
@@ -75,7 +75,7 @@ describe('StartInvestigationComponent', () => {
   it('should sort table data', async () => {
     const fixture = await renderStartInvestigation();
     const spy = spyOn((fixture.componentInstance as any).childPartsState, 'update').and.callThrough();
-    const nameHeader = await waitFor(() => screen.getByText('table.column.name'));
+    const nameHeader = await waitFor(() => screen.getByText('table.column.nameAtManufacturer'));
 
     fireEvent.click(nameHeader);
     expect(spy).toHaveBeenCalledWith({ data: [firstChild] });
