+++ cfx-frontend/src/app/modules/shared/shared.module.ts
@@ -94,6 +94,7 @@ import { InfoListComponent } from './components/info-list/info-list.component';
 import { RequestStepperComponent } from './components/request-notification/request-stepper/request-stepper.component';
 import { SupplierPartsComponent } from '@page/other-parts/presentation/supplier-parts/supplier-parts.component';
 import { OwnPartsComponent } from '@page/parts/presentation/own-parts/own-parts.component';
+import { TableColumnSettingsComponent } from './components/table-column-settings/table-column-settings.component';
 @NgModule({
   declarations: [
     ToastContainerComponent,
@@ -103,6 +104,7 @@ import { OwnPartsComponent } from '@page/parts/presentation/own-parts/own-parts.
     TextWithIconComponent,
     TableComponent,
     TableSettingsComponent,
+    TableColumnSettingsComponent,
     TooltipDirective,
     RoleDirective,
     I18nPipe,
